functionality done

+++ src/Components/ArticleComments.jsx
@@ -2,12 +2,14 @@ import React, { Component } from "react";
 import { getCommentsByArticleId, deleteCommentByCommentId } from "../api";
 import LoadingImage from "./LoadingImage";
 import CommentAdder from "./CommentAdder";
-import CommentVoter from "./CommentVoter";
+import VoteCaster from "./VoteCaster";
+import ErrorDisplay from "./ErrorDisplay";
 
 class ArticleComments extends Component {
   state = {
     comments: [],
-    isLoading: true
+    isLoading: true,
+    err: null
   };
 
   componentDidMount() {
@@ -15,9 +17,14 @@ class ArticleComments extends Component {
   }
   fetchArticleComments = () => {
     const { article_id } = this.props;
-    getCommentsByArticleId(article_id).then(data =>
-      this.setState({ comments: data, isLoading: false })
-    );
+    getCommentsByArticleId(article_id)
+      .then(data => this.setState({ comments: data, isLoading: false }))
+      .catch(({ response }) =>
+        this.setState({
+          err: { status: response.status, msg: response.data.msg },
+          isLoading: false
+        })
+      );
   };
   componentDidUpdate(prevProps, prevState) {
     if (prevState.comments.comment_id !== this.state.comments.comment_id) {
@@ -32,7 +39,9 @@ class ArticleComments extends Component {
   };
 
   handleDelete = comment_id => {
-    deleteCommentByCommentId(comment_id);
+    const { user } = this.props;
+    const { author } = this.state.comments;
+    if (user === author) deleteCommentByCommentId(comment_id);
     const filteredComments = this.state.comments.filter(
       com => com.comment_id !== comment_id
     );
@@ -40,11 +49,11 @@ class ArticleComments extends Component {
   };
 
   render() {
-    const { comments, isLoading } = this.state;
+    const { comments, isLoading, err } = this.state;
     const { article_id, user } = this.props;
 
     if (isLoading) return <LoadingImage />;
-
+    if (err) return <ErrorDisplay err={err} />;
     return (
       <div>
         <CommentAdder
@@ -58,12 +67,15 @@ class ArticleComments extends Component {
               <li key={comment.comment_id}>
                 <h6>{comment.author}</h6>
                 <p>{comment.body}</p>
-                <p>votes:{comment.votes}</p>
+
                 <p>submitted: {comment.created_at}</p>
                 <button onClick={() => this.handleDelete(comment.comment_id)}>
                   delete
                 </button>
-                <CommentVoter comment_id={comment.comment_id} />
+                <VoteCaster
+                  votes={comment.votes}
+                  comment_id={comment.comment_id}
+                />
               </li>
             );
           })}
